app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
llama-index-core==0.9.12
llama-index-vector-stores-chroma==0.9.12
llama-index-embeddings-huggingface==0.9.12
chromadb==0.4.15
sentence-transformers==2.2.2
torch==2.1.0
pydantic==2.5.0